Backend/OmniBliss/requirements.txt
asgiref==3.3.4
Django==3.2
django-countries==7.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.4
phonenumbers==8.12.21
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
